Added resetting users online state on startup Moved login and logout into their on files
